Move remaining logic to install library
