+++ src/functions/kits/install.ts
@@ -1,25 +1,32 @@
 import * as crypto from "crypto";
 import * as path from "path";
+import * as clc from "colorette";
 import * as fs from "fs-extra";
 
 import { Config } from "../../config";
 import { FirebaseError, getErrMsg } from "../../error";
 import { KitFunctionConfig, FunctionsConfig } from "../../firebaseConfig";
 import { getProjectId } from "../../projectUtils";
-import { logLabeledBullet } from "../../utils";
+import { logLabeledBullet, logLabeledSuccess, logLabeledWarning } from "../../utils";
 import {
+  addKitPrefix,
   isKitConfig,
   normalizeAndValidate,
+  validateKit,
+  validateKitInstanceId,
   ValidatedKitSingle,
   ValidatedSingle,
 } from "../projectConfig";
 import { logger } from "../../logger";
+import { confirm, input, select } from "../../prompt";
 import { spawnWithOutput, wrapSpawn } from "../../init/spawn";
 import { readTemplateSync } from "../../templates";
 import * as supported from "../../deploy/functions/runtimes/supported";
 import * as runtimes from "../../deploy/functions/runtimes";
 import * as build from "../../deploy/functions/build";
+import * as iam from "../../gcp/iam";
 import { hasProjectEnv } from "../env";
+import { RC } from "../../rc";
 import { KitInstanceEnvSeed, seedKitInstanceEnv } from "./env";
 
 export const TEMPLATES = {
@@ -65,6 +72,43 @@ export interface AddKitInstanceOptions {
 export interface AddKitInstanceResult {
   configDirPath: string;
   absConfigDirPath: string;
+}
+
+export interface InstallKitOrInstanceOptions {
+  config: Config;
+  package: string;
+  template?: TemplateType;
+  kitId?: string;
+  instanceId?: string;
+  seedEnv?: KitInstanceEnvSeed;
+  nonInteractive?: boolean;
+  project?: string;
+  projectId?: string;
+  rc?: RC;
+}
+
+export interface InstallKitOrInstanceResult {
+  action: "installedKit" | "addedInstance" | "configuredEnv";
+  kitId: string;
+  instanceId?: string;
+  sourcePath?: string;
+  configDirPath?: string;
+}
+
+export interface PromptExistingInstanceOptions {
+  project?: string;
+  projectId?: string;
+  nonInteractive?: boolean;
+}
+
+export interface ExistingKitInstallOptions {
+  config: Config;
+  project?: string;
+  projectId?: string;
+  nonInteractive?: boolean;
+  rc?: RC;
+  instanceId?: string;
+  seedEnv?: KitInstanceEnvSeed;
 }
 
 /**
@@ -212,6 +256,194 @@ export function extractExistingFunctionsInfo(
     existingCodebases,
     existingInstanceIds,
   };
+}
+
+/**
+ * Prompts the user for a kit instance ID with validation against collision with existing instance IDs and codebase names.
+ */
+export async function promptKitInstanceId(
+  baseKitId: string,
+  existingInstanceIds: Set<string>,
+  existingCodebases: Set<string>,
+  nonInteractive?: boolean,
+  customInstanceId?: string,
+): Promise<string> {
+  const instanceCollisions = new Set([...existingInstanceIds, ...existingCodebases]);
+  const defaultInstanceId = generateUniqueId(baseKitId, instanceCollisions);
+
+  if (customInstanceId) {
+    validateKitInstanceId(customInstanceId);
+    if (existingInstanceIds.has(customInstanceId)) {
+      throw new FirebaseError(
+        `functions kit instance ID must be unique across all kits, but '${customInstanceId}' was used more than once.`,
+      );
+    }
+    if (existingCodebases.has(customInstanceId)) {
+      throw new FirebaseError(
+        `functions codebase name and kit instance ID must be mutually exclusive, but '${customInstanceId}' was used as both a codebase name and a kit instance ID.`,
+      );
+    }
+    return customInstanceId;
+  }
+
+  const instanceId = await input({
+    message: "What would you like to name this instance?",
+    default: defaultInstanceId,
+    nonInteractive,
+    validate: (val: string) => {
+      try {
+        validateKitInstanceId(val);
+      } catch (err: unknown) {
+        return getErrMsg(err);
+      }
+      if (existingInstanceIds.has(val)) {
+        return `functions kit instance ID must be unique across all kits, but '${val}' was used more than once.`;
+      }
+      if (existingCodebases.has(val)) {
+        return `functions codebase name and kit instance ID must be mutually exclusive, but '${val}' was used as both a codebase name and a kit instance ID.`;
+      }
+      return true;
+    },
+  });
+
+  validateKitInstanceId(instanceId);
+  if (existingInstanceIds.has(instanceId)) {
+    throw new FirebaseError(
+      `functions kit instance ID must be unique across all kits, but '${instanceId}' was used more than once.`,
+    );
+  }
+  if (existingCodebases.has(instanceId)) {
+    throw new FirebaseError(
+      `functions codebase name and kit instance ID must be mutually exclusive, but '${instanceId}' was used as both a codebase name and a kit instance ID.`,
+    );
+  }
+
+  return instanceId;
+}
+
+/**
+ * Prompts the user for a kit ID with validation against existing kit IDs.
+ */
+export async function promptKitId(
+  packageName: string,
+  existingKitIds: Set<string>,
+  nonInteractive?: boolean,
+  customKitId?: string,
+): Promise<string> {
+  const baseKitId = sanitizePackageNameToKitName(packageName);
+  const defaultKitId = generateUniqueId(baseKitId, existingKitIds);
+
+  if (customKitId) {
+    validateKit(customKitId);
+    if (existingKitIds.has(customKitId)) {
+      throw new FirebaseError(
+        `functions.kit must be unique but '${customKitId}' was used more than once.`,
+      );
+    }
+    return customKitId;
+  }
+
+  const kitId = await input({
+    message: "What would you like to name this kit?",
+    default: defaultKitId,
+    nonInteractive,
+    validate: (val: string) => {
+      try {
+        validateKit(val);
+      } catch (err: unknown) {
+        return getErrMsg(err);
+      }
+      if (existingKitIds.has(val)) {
+        return `functions.kit must be unique but '${val}' was used more than once.`;
+      }
+      return true;
+    },
+  });
+
+  validateKit(kitId);
+  if (existingKitIds.has(kitId)) {
+    throw new FirebaseError(`functions.kit must be unique but '${kitId}' was used more than once.`);
+  }
+
+  return kitId;
+}
+
+/**
+ * Warns about third-party packages or missing shrinkwrap, and prompts for user confirmation before installation.
+ */
+export async function promptSecurityConfirmation(
+  rawPkgName: string,
+  packageName: string,
+  nonInteractive?: boolean,
+): Promise<boolean> {
+  const isThirdParty = isThirdPartyPackage(packageName);
+  if (isThirdParty) {
+    logLabeledWarning(
+      "functions",
+      `Package ${clc.bold(packageName)} is a third-party kit (outside the @firebase-functions-kits scope).`,
+    );
+  }
+
+  const hasShrinkwrap = await checkPackageHasShrinkwrap(rawPkgName);
+  if (!hasShrinkwrap) {
+    logLabeledWarning(
+      "functions",
+      `Package ${clc.bold(packageName)} does not have an npm-shrinkwrap.json file. npm-shrinkwrap guarantees that you deploy the same version of dependencies that the publisher tested against. Since this kit does not have an npm-shrinkwrap, it is possible that deploys or updates may introduce bugs or vulnerabilities in newer dependency versions that the publisher did not test against.`,
+    );
+  }
+
+  if (isThirdParty || !hasShrinkwrap) {
+    let confirmMessage: string;
+    if (isThirdParty && !hasShrinkwrap) {
+      confirmMessage = `Are you sure you want to install the third-party kit ${packageName} without locked dependencies?`;
+    } else if (isThirdParty) {
+      confirmMessage = `Are you sure you want to install the third-party kit ${packageName}?`;
+    } else {
+      confirmMessage = `Are you sure you want to install ${packageName} without locked dependencies?`;
+    }
+    const confirmInstallation = await confirm({
+      message: confirmMessage,
+      default: false,
+      nonInteractive,
+    });
+    if (!confirmInstallation) {
+      throw new FirebaseError("Installation cancelled.");
+    }
+  }
+
+  return isThirdParty;
+}
+
+/**
+ * Guides the user on configuring an existing instance for the active project.
+ */
+export async function promptExistingInstanceForProject(
+  options: PromptExistingInstanceOptions,
+  existingKit: ValidatedKitSingle,
+): Promise<string> {
+  const instanceIds: string[] = Object.keys(existingKit.instances || {});
+  if (instanceIds.length === 0) {
+    throw new FirebaseError(`Kit '${existingKit.kit}' has no instances configured.`);
+  }
+
+  let selectedInstanceId = "<instance-name>";
+  if (instanceIds.length === 1) {
+    selectedInstanceId = instanceIds[0];
+  } else if (!options.nonInteractive) {
+    selectedInstanceId = await select<string>({
+      message: "Which instance would you like to configure for this project?",
+      choices: instanceIds.map((id) => ({ name: id, value: id })),
+    });
+  }
+
+  const targetProject = getProjectId(options) || options.project || "<project-name>";
+  logLabeledBullet(
+    "functions",
+    `To create a new instance in this project, deploy the instance dedicated to this project using\n` +
+      clc.bold(`firebase deploy --only functions:${selectedInstanceId} --project ${targetProject}`),
+  );
+
+  return selectedInstanceId;
 }
 
 /**
@@ -500,4 +732,219 @@ export async function discoverKitBuild(
   };
   const runtimeDelegate = await runtimes.getRuntimeDelegate(delegateContext);
   return runtimeDelegate.discoverBuild({}, {});
+}
+
+/**
+ * Discovers kit endpoints, required APIs, and required roles, and logs a formatted report.
+ */
+export async function printKitFirstDeployReport(
+  options: { config?: Config; project?: string; projectId?: string },
+  instanceId: string,
+  absSourcePath: string,
+): Promise<void> {
+  let discoveredBuild: build.Build;
+  const prefix = addKitPrefix(instanceId);
+  try {
+    discoveredBuild = await discoverKitBuild(options, absSourcePath);
+    build.applyPrefix(discoveredBuild, prefix);
+  } catch (err: unknown) {
+    logger.debug(`Could not discover kit build for reporting: ${getErrMsg(err)}`);
+    return;
+  }
+
+  const functions = Object.keys(discoveredBuild.endpoints).sort();
+  const apis = (discoveredBuild.requiredAPIs || []).map((a) => a.api).sort();
+  const rawRoles = discoveredBuild.requiredRoles || [];
+  const roles = (await Promise.all(rawRoles.map((r) => iam.getRoleName(r)))).sort();
+
+  const printSection = (heading: string, items: string[]): void => {
+    if (items.length > 0) {
+      logLabeledBullet("functions", `${heading}\n` + items.map((item) => `- ${item}`).join("\n"));
+    }
+  };
+
+  printSection(
+    "At the first deploy, the following functions will be created in your project:",
+    functions,
+  );
+  printSection("At the first deploy, the following APIs will be enabled in your project:", apis);
+  printSection(
+    "At the first deploy, the following roles will be granted to the kit service account:",
+    roles,
+  );
+}
+
+/**
+ * Handles installation when the kit package is already present in firebase.json.
+ */
+export async function addKitInstanceOrConfigureProject(
+  options: ExistingKitInstallOptions,
+  existingKit: ValidatedKitSingle,
+  existingFunctionsInfo: ExistingFunctionsInfo,
+): Promise<InstallKitOrInstanceResult> {
+  const projectId = getProjectId(options) || options.projectId;
+  const projectAlias =
+    options.rc?.hasProjects && options.project && options.rc.hasProjectAlias(options.project)
+      ? options.project
+      : undefined;
+  const isConfiguredForProject = isKitConfiguredForProject(
+    options.config,
+    existingKit,
+    projectId,
+    projectAlias,
+  );
+
+  let action: "addInstance" | "addEnv";
+  if (!isConfiguredForProject && !options.nonInteractive) {
+    const existingInstances = Object.keys(existingKit.instances || {}).join(", ");
+    action = await select<"addInstance" | "addEnv">({
+      message: `The following instances already exist, but are not configured for this project: ${existingInstances}. What would you like to do?`,
+      choices: [
+        {
+          name: "Add an instance to the existing kit",
+          value: "addInstance",
+        },
+        {
+          name: "Configure an existing instance for this project",
+          value: "addEnv",
+        },
+      ],
+    });
+  } else {
+    if (isConfiguredForProject) {
+      logLabeledBullet(
+        "functions",
+        `This package is already installed as kit ${existingKit.kit}, creating a new instance.`,
+      );
+    }
+    action = "addInstance";
+  }
+
+  if (action === "addInstance") {
+    const instanceId = await promptKitInstanceId(
+      existingKit.kit,
+      existingFunctionsInfo.existingInstanceIds,
+      existingFunctionsInfo.existingCodebases,
+      options.nonInteractive,
+      options.instanceId,
+    );
+
+    const result = await addInstanceToKit({
+      config: options.config,
+      kitId: existingKit.kit,
+      instanceId,
+      seedEnv: options.seedEnv,
+    });
+
+    logLabeledSuccess(
+      "functions",
+      `Function kit instance ${clc.bold(instanceId)} successfully added to kit ${clc.bold(existingKit.kit)}.`,
+    );
+    await printKitFirstDeployReport(options, instanceId, options.config.path(existingKit.source));
+
+    return {
+      action: "addedInstance",
+      kitId: existingKit.kit,
+      instanceId,
+      sourcePath: existingKit.source,
+      configDirPath: result.configDirPath,
+    };
+  }
+
+  const selectedInstanceId = await promptExistingInstanceForProject(options, existingKit);
+  return {
+    action: "configuredEnv",
+    kitId: existingKit.kit,
+    instanceId: selectedInstanceId,
+  };
+}
+
+/**
+ * Orchestrates the complete kit installation flow.
+ * Installs a brand new kit (if not present) or adds a new instance to an existing kit.
+ */
+export async function installKitOrInstance(
+  options: InstallKitOrInstanceOptions,
+): Promise<InstallKitOrInstanceResult> {
+  const templateType = options.template || DEFAULT_TEMPLATE;
+  if (!(templateType in TEMPLATES)) {
+    const validTemplates = Object.keys(TEMPLATES)
+      .map((t) => `'${t}'`)
+      .join(" or ");
+    throw new FirebaseError(
+      `Invalid template '${templateType}'. Template must be ${validTemplates}.`,
+    );
+  }
+
+  const rawPkgName = options.package;
+  if (!rawPkgName) {
+    throw new FirebaseError("Set the --package option to a valid NPM package and try again.");
+  }
+
+  const { packageName } = parseNpmPackageSpecifier(rawPkgName);
+  validateNpmPackageName(packageName);
+
+  const existingFunctionsInfo = extractExistingFunctionsInfo(options.config.src.functions);
+  const existingKit = existingFunctionsInfo.existingFunctions.find(
+    (c): c is ValidatedKitSingle => isKitConfig(c) && c.sourcePackage?.name === packageName,
+  );
+
+  if (existingKit) {
+    return addKitInstanceOrConfigureProject(options, existingKit, existingFunctionsInfo);
+  }
+
+  const isThirdParty = await promptSecurityConfirmation(
+    rawPkgName,
+    packageName,
+    options.nonInteractive,
+  );
+
+  const kitId = await promptKitId(
+    packageName,
+    existingFunctionsInfo.existingKitIds,
+    options.nonInteractive,
+    options.kitId,
+  );
+
+  const instanceId = await promptKitInstanceId(
+    kitId,
+    existingFunctionsInfo.existingInstanceIds,
+    existingFunctionsInfo.existingCodebases,
+    options.nonInteractive,
+    options.instanceId,
+  );
+
+  const { sourcePath, configDirPath, absSourcePath, absConfigDirPath } = await scaffoldKitFiles(
+    options.config,
+    kitId,
+    instanceId,
+    packageName,
+    templateType,
+  );
+
+  if (options.seedEnv?.envs && Object.keys(options.seedEnv.envs).length > 0) {
+    seedKitInstanceEnv({
+      configDir: absConfigDirPath,
+      functionsSource: absSourcePath,
+      projectDir: options.config.projectDir,
+      projectId: options.seedEnv.projectId,
+      projectAlias: options.seedEnv.projectAlias,
+      envs: options.seedEnv.envs,
+    });
+  }
+
+  await buildAndInstallKit(absSourcePath, rawPkgName, isThirdParty);
+
+  addKitToConfig(options.config, kitId, instanceId, packageName, sourcePath, configDirPath);
+
+  logLabeledSuccess("functions", `Function kit ${clc.bold(kitId)} successfully installed.`);
+  await printKitFirstDeployReport(options, instanceId, absSourcePath);
+
+  return {
+    action: "installedKit",
+    kitId,
+    instanceId,
+    sourcePath,
+    configDirPath,
+  };
 }
